reorganize gen-types files
