+++ codegen/gen_ts_types/javap.ts
@@ -1,0 +1,6 @@
+// @ts-ignore
+import { execSync } from "child_process";
+
+export const javap = (file: string): string => {
+	return execSync(`javap -public ${file}`).toString();
+};
