+++ codegen/gen_ts_types/test_files/Example.class
